% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_to_rast.R
\name{feat_to_rast}
\alias{feat_to_rast}
\title{Load an Arc raster dataset (from .gdb) as a terra SpatRast}
\usage{
feat_to_rast(infile)
}
\arguments{
\item{infile}{Path to the input raster}
}
\value{
A SpatRast of the infile
}
\description{
This function loads a raster dataset from the ESRI environment,
from a .gdb, and converts the raster into a
SpatRast from the terra package. CRS information is maintained.
}
\examples{
my_raster<- feat_to_rast("./geodatabase_path/geodatabase.gdb/raster")
}
