% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_to_raster.R
\name{feat_to_raster}
\alias{feat_to_raster}
\title{Load an Arc raster dataset (from .gdb) as a RasterLayer}
\usage{
feat_to_raster(infile)
}
\arguments{
\item{infile}{Path to the input raster}
}
\value{
A RasterLayer of the infile
}
\description{
This function loads a raster dataset from the ESRI environment,
from a .gdb, and converts the raster into a
RasterLayer from the raster package. CRS information is maintained.
}
\examples{
my_raster<- feat_to_raster("./geodatabase_path/geodatabase.gdb/raster")
}
