% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_feat.R
\name{write_to_feat}
\alias{write_to_feat}
\title{Write a terra SpatVect or sp object to an Arc feature (in a .gdb or as a .shp)}
\usage{
write_to_feat(infile, file_path)
}
\arguments{
\item{infile}{SpatVect or sp object in your environment}

\item{file_path}{Full file path where you want to save - either a folder for a .shp or a .gdb (can be new or existing), and the file name (no extension necessary)}
}
\value{
The object will be saved where specified
}
\description{
This function writes a terra SpatVect or sp object as an Arc feature,
either as a standalone .shp or a feature in a .gdb.
A new .gdb can also be made during this step.
CRS information may be altered.
This function will work for SpatVect objects or Spatial objects (spatial polygons/points data frames) and potentially other file types. If you contact me I can try to add support for more types.
}
\examples{
save object as .shp:  write_vect("./desired_file_path/feature")
save object as feature in existing .gdb:  write_vect("./geodatabase_path/geodatabase.gdb/feature")
save object as feature in new .gdb:  write_vect("./geodatabase_path/new_geodatabase.gdb/feature")

}
